OD-10877 Services app, contact merge: tag relation stubs should be removed before regular transaction processing (to avoid exception 'Can not find corresponded record Taggable')
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@51453 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ services/src/main/java/ish/oncourse/webservices/replication/services/TagRelationDeduper.java
@@ -32,15 +32,17 @@ public class TagRelationDeduper {
     public void dedupe() {
         List<Taggable> entityTaggables = getAllOf(entity);
 
-        Taggable headTaggable = defineHeadTaggable(entityTaggables);
-        List<Taggable> taggableDuplicates = defineDuplicates(entityTaggables);
+        if (entityTaggables.size() > 1) {
+            Taggable headTaggable = defineHeadTaggable(entityTaggables);
+            List<Taggable> taggableDuplicates = defineDuplicates(entityTaggables);
 
-        relinkTaggableRelations(taggableDuplicates, headTaggable);
-        List<Taggable> unusuableTaggables = taggableDuplicates;
-        List<TaggableTag> unusuableTaggableTags = dedupeTaggableTags(headTaggable);
+            relinkTaggableRelations(taggableDuplicates, headTaggable);
+            List<Taggable> unusuableTaggables = taggableDuplicates;
+            List<TaggableTag> unusuableTaggableTags = dedupeTaggableTags(headTaggable);
 
-        context.deleteObjects(unusuableTaggableTags);
-        context.deleteObjects(unusuableTaggables);
+            context.deleteObjects(unusuableTaggableTags);
+            context.deleteObjects(unusuableTaggables);
+        }
     }
 
     /**
